+++ src/js/Helper/Import/CsvConversionHelper.js
@@ -92,7 +92,9 @@ export default class ImportCsvConversionHelper {
             db        = [],
             firstLine = Number.parseInt(0 + options.firstLine);
 
-        if(data[firstLine].length < mapping.length) throw new Error('CSV file can not be mapped');
+        if(!Array.isArray(data) || data.length < firstLine+1) throw new Error('Import file is empty');
+        if(!mapping || data[firstLine].length < mapping.length) throw new Error('CSV file can not be mapped');
+
         for(let i = firstLine; i < data.length; i++) {
             let line   = data[i],
                 object = {};
@@ -415,6 +417,11 @@ export default class ImportCsvConversionHelper {
                 firstLine: 1,
                 db       : 'passwords',
                 mapping  : ['label', 'url', '', 'username', 'password', 'notes']
+            },
+            safeincloud : {
+                firstLine: 1,
+                db       : 'passwords',
+                mapping  : ['label', 'username', 'password', 'url', 'notes']
             }
         };