+++ src/lib/Command/BackupDeleteCommand.php
@@ -7,19 +7,21 @@
 
 namespace OCA\Passwords\Command;
 
+use Exception;
 use OCA\Passwords\Services\BackupService;
-use OCP\Util;
+use OCP\Files\NotFoundException;
+use OCP\Files\NotPermittedException;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Class BackupCreateCommand
+ * Class BackupDeleteCommand
  *
  * @package OCA\Passwords\Command
  */
-class BackupCreateCommand extends Command {
+class BackupDeleteCommand extends Command {
 
     /**
      * @var BackupService
@@ -41,9 +43,9 @@ class BackupCreateCommand extends Command {
      *
      */
     protected function configure(): void {
-        $this->setName('passwords:backup:create')
-             ->addArgument('name', InputArgument::OPTIONAL, 'The name of the backup')
-             ->setDescription('Create a new backup of the password database');
+        $this->setName('passwords:backup:delete')
+             ->addArgument('backup', InputArgument::REQUIRED, 'The backup to delete')
+             ->setDescription('Delete a manually created backup file');
     }
 
     /**
@@ -51,19 +53,19 @@ class BackupCreateCommand extends Command {
      * @param OutputInterface $output
      *
      * @return int|null|void
-     * @throws \OCP\Files\NotPermittedException
-     * @throws \OCP\Files\NotFoundException
+     * @throws NotPermittedException
+     * @throws Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output): void {
-        $argName = null;
-        if($input->hasArgument('name')) {
-            $argName = preg_replace('/[^\w\-\.]/', '', $input->getArgument('name'));
-            if(empty($argName)) $argName = null;
+        $backup  = $input->getArgument('backup');
+        $backups = $this->backupService->getBackups('backups');
+
+        if(isset($backups[ $backup ])) {
+            $backups[ $backup ]->delete();
+
+            $output->writeln('Deleted: '.$backup);
+        } else {
+            throw new NotFoundException();
         }
-
-        $backup = $this->backupService->createBackup($argName);
-        $info = $this->backupService->getBackupInfo($backup);
-
-        $output->writeln(sprintf('Created new backup: %s, %s %s', $info['label'], $info['size'], $info['format']));
     }
 }